fixed build regression
git-svn-id: http://svn.code.sf.net/p/qmmp-dev/code/trunk/qmmp@9244 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/plugins/Input/vorbis/decoder_vorbis.cpp
@@ -122,7 +122,7 @@ bool DecoderVorbis::initialize()
     {
         freq = ogginfo->rate;
         chan = ogginfo->channels;
-        setProperty(Qmmp::BITRATE, ogginfo->bitrate_nominal / 1000);
+        setProperty(Qmmp::BITRATE, int(ogginfo->bitrate_nominal / 1000));
     }
 
     ChannelMap chmap = findChannelMap(chan);
